Fix xross connection tests

+++ controlplane/pkg/local/monitor_service.go
@@ -53,6 +53,8 @@ func (srv *monitorService) Close(ctx context.Context, connection *connection.Con
 	// Pass model connection with context
 	ctx = common.WithMonitorServer(ctx, srv.monitor)
 	conn, err := ProcessClose(ctx, connection)
-	srv.monitor.Delete(ctx, connection)
+	if conn != nil {
+		srv.monitor.Delete(ctx, connection)
+	}
 	return conn, err
 }
